Increased parallelism, buffer size and max spout pending

+++ src/main/java/fogetti/phish/storm/integration/PhishTopologyRemoteRunner.java
@@ -13,11 +13,11 @@ public class PhishTopologyRemoteRunner {
 		config.setNumWorkers(20);
 		config.setMessageTimeoutSecs(30);
 		config.put(Config.TOPOLOGY_EXECUTOR_RECEIVE_BUFFER_SIZE,
-	               new Integer(128));
+	               new Integer(524288));
 	    config.put(Config.TOPOLOGY_EXECUTOR_SEND_BUFFER_SIZE,
-	               new Integer(128));
+	               new Integer(524288));
 	    config.put(Config.TOPOLOGY_TRANSFER_BUFFER_SIZE,
-	               new Integer(128));
+	               new Integer(524288));
 		
 		StormSubmitter.submitTopology("phish-storm-topology", config, topology);
 	}
